+++ front/components/assistant_builder/AssistantBuilderTablesModal.tsx
@@ -1,13 +1,4 @@
-import {
-  Button,
-  GlobeAltIcon,
-  Item,
-  Modal,
-  Page,
-  Searchbar,
-  ServerIcon,
-  Spinner,
-} from "@dust-tt/sparkle";
+import { Modal, Spinner } from "@dust-tt/sparkle";
 import type {
   ContentNode,
   CoreAPITable,
@@ -17,24 +8,22 @@ import type {
 import {
   canContainStructuredData,
   getMicrosoftSheetContentNodeInternalIdFromTableId,
+  isFolder,
 } from "@dust-tt/types";
 import {
   getGoogleSheetContentNodeInternalIdFromTableId,
   getNotionDatabaseContentNodeInternalIdFromTableId,
   getTableIdForContentNode,
 } from "@dust-tt/types";
-import { Transition } from "@headlessui/react";
 import * as React from "react";
 import { useEffect, useMemo, useState } from "react";
 
 import { AssistantBuilderContext } from "@app/components/assistant_builder/AssistantBuilderContext";
+import PickDataSourceForTable from "@app/components/assistant_builder/PickDataSourceForTable";
+import { PickTableInFolder } from "@app/components/assistant_builder/PickTableInFolder";
+import { PickTablesManaged } from "@app/components/assistant_builder/PickTablesManaged";
 import type { AssistantBuilderTableConfiguration } from "@app/components/assistant_builder/types";
-import DataSourceResourceSelectorTree from "@app/components/DataSourceResourceSelectorTree";
-import { orderDatasourceByImportance } from "@app/lib/assistant";
-import { CONNECTOR_CONFIGURATIONS } from "@app/lib/connector_providers";
-import { getDisplayNameForDataSource } from "@app/lib/data_sources";
-import { useDataSourceNodes, useTables } from "@app/lib/swr";
-import { compareForFuzzySort, subFilter } from "@app/lib/utils";
+import { useDataSourceNodes } from "@app/lib/swr";
 
 export default function AssistantBuilderTablesModal({
   isOpen,
@@ -172,14 +161,14 @@ export default function AssistantBuilderTablesModal({
         {!selectedManagedTables ? (
           <Spinner />
         ) : !selectedDataSource ? (
-          <PickDataSource
+          <PickDataSourceForTable
             dataSources={supportedDataSources}
             onPick={(ds: DataSourceType) => {
               setSelectedDataSource(ds);
             }}
           />
-        ) : !selectedDataSource.connectorId ? (
-          <PickTable
+        ) : isFolder(selectedDataSource) ? (
+          <PickTableInFolder
             owner={owner}
             dataSource={selectedDataSource}
             onPick={(table: CoreAPITable) => {
@@ -261,181 +250,3 @@ export default function AssistantBuilderTablesModal({
     </Modal>
   );
 }
-
-function PickDataSource({
-  dataSources,
-  onPick,
-}: {
-  dataSources: DataSourceType[];
-  onPick: (dataSource: DataSourceType) => void;
-}) {
-  const [query, setQuery] = useState<string>("");
-
-  const filtered = dataSources.filter((ds) => {
-    return subFilter(query.toLowerCase(), ds.name.toLowerCase());
-  });
-
-  return (
-    <Transition show={true} className="mx-auto max-w-6xl">
-      <Page>
-        <Page.Header title="Select a Table in" icon={ServerIcon} />
-        <Searchbar
-          name="search"
-          onChange={setQuery}
-          value={query}
-          placeholder="Search..."
-        />
-        {orderDatasourceByImportance(filtered).map((ds) => (
-          <Item.Navigation
-            label={getDisplayNameForDataSource(ds)}
-            icon={
-              ds.connectorProvider
-                ? CONNECTOR_CONFIGURATIONS[ds.connectorProvider].logoComponent
-                : GlobeAltIcon
-            }
-            key={ds.id}
-            onClick={() => {
-              onPick(ds);
-            }}
-          />
-        ))}
-      </Page>
-    </Transition>
-  );
-}
-
-const PickTable = ({
-  owner,
-  dataSource,
-  onPick,
-  onBack,
-  tablesQueryConfiguration,
-}: {
-  owner: WorkspaceType;
-  dataSource: DataSourceType;
-  onPick: (table: CoreAPITable) => void;
-  onBack?: () => void;
-  tablesQueryConfiguration: Record<string, AssistantBuilderTableConfiguration>;
-}) => {
-  const { tables } = useTables({
-    workspaceId: owner.sId,
-    dataSourceName: dataSource.name,
-  });
-  const [query, setQuery] = useState<string>("");
-
-  const tablesToDisplay = tables.filter(
-    (t) =>
-      !tablesQueryConfiguration?.[
-        `${owner.sId}/${dataSource.name}/${t.table_id}`
-      ]
-  );
-  const filtered = useMemo(
-    () =>
-      tablesToDisplay.filter((t) => {
-        return subFilter(query.toLowerCase(), t.name.toLowerCase());
-      }),
-    [query, tablesToDisplay]
-  );
-
-  const isAllSelected = !!tables.length && !tablesToDisplay.length;
-
-  return (
-    <Transition show={true} className="mx-auto max-w-6xl">
-      <Page>
-        <Page.Header title="Select a Table" icon={ServerIcon} />
-        {isAllSelected && (
-          <div className="flex h-full w-full flex-col">
-            <div className="text-gray-500">
-              All tables from this DataSource are already selected.
-            </div>
-          </div>
-        )}
-
-        {tables.length === 0 && (
-          <div className="flex h-full w-full flex-col">
-            <div className="text-gray-500">
-              No tables found in this Data Source.
-            </div>
-          </div>
-        )}
-
-        {!!tablesToDisplay.length && (
-          <>
-            <Searchbar
-              name="search"
-              onChange={setQuery}
-              value={query}
-              placeholder="Search..."
-            />
-            {filtered
-              .sort((a, b) => compareForFuzzySort(query, a.name, b.name))
-              .map((table) => {
-                return (
-                  <Item.Navigation
-                    label={table.name}
-                    icon={
-                      dataSource.connectorProvider
-                        ? CONNECTOR_CONFIGURATIONS[dataSource.connectorProvider]
-                            .logoComponent
-                        : ServerIcon
-                    }
-                    key={`${table.data_source_id}/${table.table_id}`}
-                    onClick={() => {
-                      onPick(table);
-                    }}
-                  />
-                );
-              })}
-          </>
-        )}
-
-        <div className="flex pt-8">
-          <Button label="Back" onClick={onBack} variant="secondary" />
-        </div>
-      </Page>
-    </Transition>
-  );
-};
-
-const PickTablesManaged = ({
-  owner,
-  dataSource,
-  onSelectionChange,
-  selectedNodes,
-  parentsById,
-}: {
-  owner: WorkspaceType;
-  dataSource: DataSourceType;
-  onSelectionChange: (
-    resource: ContentNode,
-    parents: string[],
-    selected: boolean
-  ) => void;
-  onBack?: () => void;
-  selectedNodes: ContentNode[];
-  parentsById: Record<string, Set<string>>;
-}) => {
-  return (
-    <Transition show={true} className="mx-auto max-w-6xl">
-      <Page>
-        <Page.Header title="Select a Table" icon={ServerIcon} />
-        <DataSourceResourceSelectorTree
-          owner={owner}
-          dataSource={dataSource}
-          showExpand={true}
-          selectedResourceIds={
-            selectedNodes
-              ? [...new Set(selectedNodes.map((n) => n.internalId))]
-              : []
-          }
-          selectedParents={[
-            ...new Set(Object.values(parentsById).flatMap((c) => [...c])),
-          ]}
-          filterPermission="read"
-          viewType={"tables"}
-          onSelectChange={onSelectionChange}
-        />
-      </Page>
-    </Transition>
-  );
-};
