+++ front/components/assistant_builder/DataSourceResourceSelector.tsx
@@ -1,0 +1,108 @@
+import { Page, SliderToggle } from "@dust-tt/sparkle";
+import type {
+  ConnectorProvider,
+  ContentNode,
+  DataSourceType,
+  WorkspaceType,
+} from "@dust-tt/types";
+import { Transition } from "@headlessui/react";
+
+import { CONNECTOR_PROVIDER_TO_RESOURCE_NAME } from "@app/components/assistant_builder/shared";
+import DataSourceResourceSelectorTree from "@app/components/DataSourceResourceSelectorTree";
+import { useParentResourcesById } from "@app/hooks/useParentResourcesById";
+import { CONNECTOR_CONFIGURATIONS } from "@app/lib/connector_providers";
+import { getDisplayNameForDataSource } from "@app/lib/data_sources";
+
+export default function DataSourceResourceSelector({
+  dataSource,
+  owner,
+  selectedResources,
+  isSelectAll,
+  onSelectChange,
+  toggleSelectAll,
+}: {
+  dataSource: DataSourceType | null;
+  owner: WorkspaceType;
+  selectedResources: ContentNode[];
+  isSelectAll: boolean;
+  onSelectChange: (resource: ContentNode, selected: boolean) => void;
+  toggleSelectAll: () => void;
+}) {
+  const { parentsById, setParentsById } = useParentResourcesById({
+    owner,
+    dataSource,
+    selectedResources,
+  });
+
+  const selectedParents = [
+    ...new Set(Object.values(parentsById).flatMap((c) => [...c])),
+  ];
+
+  return (
+    <Transition show={!!dataSource} className="mx-auto max-w-6xl pb-8">
+      <Page>
+        <Page.Header
+          title={`Select Data Sources in ${
+            dataSource ? getDisplayNameForDataSource(dataSource) : null
+          }`}
+          icon={
+            CONNECTOR_CONFIGURATIONS[
+              dataSource?.connectorProvider as ConnectorProvider
+            ]?.logoComponent
+          }
+          description="Select the files and folders that will be used by the assistant as a source for its answers."
+        />
+        {dataSource && (
+          <div className="flex flex-row gap-32">
+            <div className="flex-1">
+              <div className="flex gap-4 pb-8 text-lg font-semibold text-element-900">
+                Select all
+                <SliderToggle
+                  selected={isSelectAll}
+                  onClick={() => {
+                    toggleSelectAll();
+                    setParentsById({});
+                  }}
+                  size="xs"
+                />
+              </div>
+              <div className="flex flex-row pb-4 text-lg font-semibold text-element-900">
+                <div>
+                  Select from available{" "}
+                  {CONNECTOR_PROVIDER_TO_RESOURCE_NAME[
+                    dataSource.connectorProvider as ConnectorProvider
+                  ]?.plural ?? "resources"}
+                  :
+                </div>
+              </div>
+              <DataSourceResourceSelectorTree
+                owner={owner}
+                dataSource={dataSource}
+                showExpand={
+                  CONNECTOR_CONFIGURATIONS[
+                    dataSource.connectorProvider as ConnectorProvider
+                  ]?.isNested
+                }
+                selectedResourceIds={selectedResources.map((r) => r.internalId)}
+                selectedParents={selectedParents}
+                onSelectChange={(node, parents, selected) => {
+                  setParentsById((parentsById) => {
+                    const newParentsById = { ...parentsById };
+                    if (selected) {
+                      newParentsById[node.internalId] = new Set(parents);
+                    } else {
+                      delete newParentsById[node.internalId];
+                    }
+                    return newParentsById;
+                  });
+                  onSelectChange(node, selected);
+                }}
+                parentIsSelected={isSelectAll}
+              />
+            </div>
+          </div>
+        )}
+      </Page>
+    </Transition>
+  );
+}
